backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
nltk
python-multipart
spotipy
python-dotenv
lyricsgenius
beautifulsoup4
